fix: update test imports after component reorganization
Tests added in #155 referenced flat component paths that were moved
to feature subdirectories in #154.
